Normalize CRS handling
