+++ pygeoapi/provider/sensorthings_edr.py
@@ -29,6 +29,7 @@
 
 import logging
 
+from pygeoapi.crs import DEFAULT_CRS
 from pygeoapi.provider.base import ProviderNoDataError
 from pygeoapi.provider.base_edr import BaseEDRProvider
 from pygeoapi.provider.sensorthings import SensorThingsProvider
@@ -39,7 +40,7 @@ GEOGRAPHIC_CRS = {
     'coordinates': ['x', 'y'],
     'system': {
         'type': 'GeographicCRS',
-        'id': 'http://www.opengis.net/def/crs/OGC/1.3/CRS84'
+        'id': DEFAULT_CRS
     }
 }
